Allow dpi values to be doubles. This fixes unprecise scale bars in composer
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@9298 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/qgsscalecalculator.cpp
@@ -23,18 +23,18 @@
 #include "qgsrect.h"
 #include "qgsscalecalculator.h"
 
-QgsScaleCalculator::QgsScaleCalculator( int dpi, QGis::units mapUnits )
+QgsScaleCalculator::QgsScaleCalculator( double dpi, QGis::units mapUnits )
     : mDpi( dpi ), mMapUnits( mapUnits )
 {}
 
 QgsScaleCalculator::~QgsScaleCalculator()
 {}
 
-void QgsScaleCalculator::setDpi( int dpi )
+void QgsScaleCalculator::setDpi( double dpi )
 {
   mDpi = dpi;
 }
-int QgsScaleCalculator::dpi()
+double QgsScaleCalculator::dpi()
 {
   return mDpi;
 }
